django/requirements.txt
Django == 5.0
psycopg[binary,pool] == 3.1.19
djangorestframework == 3.15.1
tzdata == 2024.1
#daphne == 4.0.0
#django-allauth == 0.63.3
#django-rest-auth == 0.9.5
argon2-cffi == 23.1.0
django-cors-headers == 4.3.1
uvicorn == 0.30.1
websockets == 12.0
PyJWT == 2.8.0
requests == 2.32.3
channels[daphne] == 4.1.0